refactor: ClientConnection repalced with Producer, Client extracted
