+++ src/Producer.js
@@ -1,6 +1,6 @@
 const RTC_CONFIG =  null;
 
-class ClientConnection {
+class Producer {
   constructor(socket, id) {
     this.id = id;
     this.serverConnection = new RTCPeerConnection(RTC_CONFIG);
@@ -9,6 +9,7 @@ class ClientConnection {
     this.requestUserMedia();
     this.mediaStream = new MediaStream();
     this.mediaTracks = {};
+    console.log('Producer constructed');
   }
 
   getMediaStream() {
@@ -16,6 +17,7 @@ class ClientConnection {
   }
 
   connect() {
+    console.log('producer.connect')
     this.socket.open();
   }
 
@@ -126,6 +128,7 @@ class ClientConnection {
     this.serverConnection.chatChannel.onmessage = callback.bind(this);
   }
 
+  // for dev only; remove later
   sendMessage() {
     console.log(this.serverConnection.chatChannel.readyState);
     if (this.serverConnection.chatChannel.readyState === 'open') {
@@ -136,4 +139,4 @@ class ClientConnection {
   }
 }
 
-export default ClientConnection;
+export default Producer;
